FTP Assignment 2 code
This is Clay's interpretation of assignment 2. It gives the same output as seen on the pdf for assignment 2. The only error right now, is that it adds an extra space for some reason in between "started." and "Habit" (ie ... "started.  Habit", instead of "started. Habit") in the pdf's method 10 example.