Clay's currently working Steg.py Program
I think I have it working. It works with the examples on moodle and I messed around with it a little too. An example run could be "python3 Steg_Program.py -r -b -o1024 -wstegged-bit.bmp > the_output.bmp". This is the first example from moodle. Just click on the "this" to download it on moodle.